Split Entity and ComponentHandle into their own headers
